fix: fix style and delete old code
