+++ project-management/public/locales/ru/translation.json
@@ -82,5 +82,8 @@
     "valid3": "Пароль должен содержать только буквы и цифры и быть более 6 символов",
     "regSucces": "Регистрация прошла успешно",
     "regBtn": "Отправить"
+  },
+  "boards": {
+    "title": "Ваши доски"
   }
 }
