+++ project-management/src/pages/Board/Board.scss
@@ -1,41 +1,41 @@
-.boards {
-    padding-left: 30px;
-    padding-right: 30px;
-    &__title {
-        font-size: 40px;
-        padding-left: 200px;
-    }
-}
-.board-container {
-    padding-left: 200px;
-    padding-right: 200px;
-    display: flex;
-    gap: 15px;
-    flex-wrap: wrap;
-}
 .board {
-    border: 1.5px solid rgba(18, 0, 97, 0.171);
-    border-radius: 11px;
-    //flex-basis: 400px;
-    width: 240px;
-    height: 180px;
-    background-color: rgba(18, 0, 97, 0.171);
-    padding-left: 25px;
-    padding-right: 25px;
-    &__title {
-        font-size: 25px;
-        color: aliceblue;
-        word-wrap: break-word;
+  position: relative;
+  border: 0.15rem solid rgba(18, 0, 97, 0.171);
+  border-radius: 1.1rem;
+  width: 24rem;
+  height: 18rem;
+  background-color: rgba(18, 0, 97, 0.171);
+  padding: 2.5rem;
+  cursor: pointer;
+
+  &__title {
+    font-family: 'Circular-Book';
+    font-size: 2.5rem;
+    color: aliceblue;
+    word-wrap: break-word;
+  }
+  &__description {
+    font-family: 'Circular-Book';
+    font-size: 2rem;
+    color: rgb(193, 215, 235);
+    max-width: 18.8rem;
+    word-wrap: break-word;
+  }
+  &__trash {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    right: 1rem;
+    top: 1rem;
+    width: 2.6rem;
+    height: 2.6rem;
+    background: transparent;
+    transition: all 0.3s;
+    border-radius: 50%;
+    &:hover {
+      transition: all 0.3s;
+      background: rgb(255 255 255 / 49%);
     }
-    &__description {
-        font-size: 20px;
-        color: rgb(193, 215, 235);
-        max-width: 188px;
-        word-wrap: break-word;
-    }
-}
-.board__trash {
-    position: relative;
-    right: -98%;
-    top: 10px;
+  }
 }
